# swagger/template.yml
swagger: "2.0"
info:
  title: edu_forms API
  description: Multi-tenant educational assessment platform
  version: "1.0.0"
host: "127.0.0.1:5000"
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT access token. Example: Bearer eyJhbG..."
  WorkspaceId:
    type: apiKey
    name: X-Workspace-Id
    in: header
    description: Active workspace context (not in JWT)
...
paths:
...
  /tests:
    post:
      tags: [Tests]
      summary: Create test (exam)
      description: |
        Step 1 exam creation payload from UI.
        Scheduling, publish timing, and advanced settings are configured in later steps.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [name, subject_id]
            properties:
              name: { type: string, example: "Midterm Exam" }
              description: { type: string }
              subject_id: { type: integer, example: 1 }
              duration_minutes:
                type: integer
                default: 30
                example: 30
                description: Exam duration in minutes (default 30)
              total_score:
                type: number
                default: 100
                example: 100
                description: Maximum score (default 100)
              passing_score:
                type: number
                default: 50
                example: 50
                description: Passing score (default 50)
              auto_distribute_scores: { type: boolean, example: false }
      responses:
        201:
          description: Test created
          schema:
            type: object
            properties:
              message: { type: string, example: "Test created successfully" }
              test:
                type: object
                properties:
                  test_id: { type: integer }
                  name: { type: string }
                  description: { type: string }
                  subject_id: { type: integer }
                  subject_name: { type: string }
                  duration_minutes: { type: integer }
                  total_score: { type: number }
                  passing_score: { type: number }
                  auto_distribute_scores: { type: boolean }
                  status: { type: string, example: DRAFT }
                  slug: { type: string }
                  created_at: { type: string, format: date-time }
                  test_link:
                    type: string
                    example: "http://localhost:5173/tests/15"
                    description: |
                      Shareable student access URL (derived from FRONTEND_BASE_URL + test_id).
                      Not stored in the database.
  /tests/my:
    get:
      tags: [Tests]
      summary: My tests (current membership)
      security: [{ Bearer: [], WorkspaceId: [] }]
      responses:
        200:
          description: List created tests
  /tests/{test_id}:
    get:
      tags: [Tests]
      summary: Get test details with snapshot questions
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Test detail
    patch:
      tags: [Tests]
      summary: Update test settings
      description: |
        Editing is allowed for:
        - DRAFT tests
        - SCHEDULED tests only when scheduled_publish_at is at least 30 minutes away.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              name: { type: string }
              slug: { type: string }
              description: { type: string }
              total_score: { type: number }
              passing_score: { type: number }
              settings_config:
                type: object
                description: |
                  Exam control settings.
                  Proctoring is active server-side via proctoring.enabled.
                  Other keys are stored for exam runtime/front-end controls.
                properties:
                  proctoring:
                    type: object
                    properties:
                      enabled: { type: boolean }
                  attempt_settings:
                    type: object
                    properties:
                      max_attempts: { type: integer, minimum: 1, default: 1 }
                  navigation_settings:
                    type: object
                    properties:
                      sequential_navigation: { type: boolean }
                      allow_back_navigation: { type: boolean }
                  answer_rules:
                    type: object
                    properties:
                      require_answer_all: { type: boolean }
                      allow_skip_questions: { type: boolean }
                  display_settings:
                    type: object
                    properties:
                      shuffle_questions: { type: boolean }
                      shuffle_choices: { type: boolean }
                  review_settings:
                    type: object
                    properties:
                      allow_review_after_grading:
                        type: boolean
                        description: "When enabled, students may review their graded attempts, including their answers, correct answers, explanations, and final score."
              availability_time_mode: { type: string, enum: [SCHEDULED, FLEXIBLE] }
              starts_at: { type: string, format: date-time }
              duration_minutes: { type: integer }
              entry_window_minutes: { type: integer }
      responses:
        200:
          description: Test updated (excludes published_at, closed_at, archived_at)
        409:
          description: Scheduled exam time conflict (student or teacher)
          schema:
            type: object
            properties:
              error:
                type: string
                enum: [SCHEDULE_CONFLICT, TEACHER_SCHEDULE_CONFLICT]
    delete:
      tags: [Tests]
      summary: Delete test (creator only)
      description: |
        Permanently deletes a test and its questions, assignments, and related data.
        Only the membership that created the test may delete it.
        Fails if any student attempts exist for this test.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Test deleted
        403:
          description: Not the test creator
        422:
          description: Test has student attempts
  /tests/{test_id}/questions:
    post:
      tags: [Tests]
      summary: Add bank questions to test as snapshots
      description: Snapshot is created immediately when questions are added to test.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [question_ids]
            properties:
              question_ids:
                type: array
                items: { type: integer }
              source_type:
                type: string
                enum: [AI, QUESTION_BANK, RANDOM_FROM_BANK, MANUAL, IMPORT]
                default: QUESTION_BANK
      responses:
        201:
          description: Questions added
  /tests/{test_id}/questions/manual:
    post:
      tags: [Tests]
      summary: Add manual questions to exam
      consumes: [application/json]
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [questions]
            properties:
              questions:
                type: array
                items:
                  type: object
                  required: [type_code, body, topic_id]
                  properties:
                    type_code: { type: string, enum: [MCQ, TRUE_FALSE, MULTI_SELECT, ESSAY] }
                    body: { type: string }
                    image_path:
                      type: string
                      description: Optional image path from POST /uploads/images
                    explanation: { type: string }
                    points: { type: number }
                    difficulty: { type: string, enum: [EASY, MEDIUM, HARD] }
                    topic_id:
                      type: integer
                      description: Required. Must belong to the same subject as the test.
                    choices: { type: array, items: { type: object } }
      responses:
        201:
          description: Manual questions added
  /tests/{test_id}/questions/import-csv:
    post:
      tags: [Tests]
      summary: Import exam questions from CSV
      description: |
        Upload a teacher-friendly CSV (no JSON in cells). Download the template from
        GET /templates/exam-questions-csv.

        **Columns:** Question Type, Question, Explanation, Difficulty, Points, Topic ID,
        Choice A–F, Correct Answers (letters A–F, e.g. `B` or `ABD`), image_url (optional).

        **image_url (optional):** relative `questions/...` path or full URL from
        `POST /uploads/images` (`image_url` response field). Leave empty for no image.
        Older CSV files without this column remain supported.

        **Supported types:** MCQ, TRUE_FALSE, MULTI_SELECT, ESSAY.

        Partial import: valid rows are created; invalid rows are returned in `failed_rows`.
        Legacy format (`type_code`, `body`, `choices` as JSON) is still accepted.
      consumes: [multipart/form-data]
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - name: csv_file
          in: formData
          type: file
          required: true
      responses:
        201:
          description: CSV questions imported (possibly with row failures)
          schema:
            type: object
            properties:
              message: { type: string }
              count: { type: integer }
              questions:
                type: array
                items: { type: object }
              failed_count: { type: integer }
              failed_rows:
                type: array
                items:
                  type: object
                  properties:
                    row: { type: integer }
                    error: { type: string }
...
  /tests/{test_id}/questions/from-bank:
    post:
      tags: [Tests]
      summary: Add selected questions from a question bank
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [bank_id, question_ids]
            properties:
              bank_id: { type: integer }
              question_ids: { type: array, items: { type: integer } }
      responses:
        201:
          description: Question bank selection added
  /tests/{test_id}/questions/random-from-banks:
    post:
      tags: [Tests]
      summary: Generate exam questions from blueprint (random-from-banks)
      description: |
        Exam Blueprint Generator. Each bank defines question_count, topic percentages (must sum to 100),
        and per-topic difficulty_distribution (easy + medium + hard = 100).
        Questions are randomly selected per bank/topic/difficulty and snapshotted into the test.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [banks]
            properties:
              banks:
                type: array
                items:
                  type: object
                  required: [bank_id, question_count, topics]
                  properties:
                    bank_id: { type: integer, example: 3 }
                    question_count: { type: integer, example: 20 }
                    topics:
                      type: array
                      items:
                        type: object
                        required: [topic_id, percentage, difficulty_distribution]
                        properties:
                          topic_id: { type: integer, example: 11 }
                          percentage: { type: integer, example: 30 }
                          difficulty_distribution:
                            type: object
                            required: [easy, medium, hard]
                            properties:
                              easy: { type: integer, example: 20 }
                              medium: { type: integer, example: 50 }
                              hard: { type: integer, example: 30 }
      responses:
        201:
          description: Blueprint generated successfully
          schema:
            type: object
            properties:
              message: { type: string, example: "Blueprint generated successfully" }
              count: { type: integer }
              summary:
                type: array
                items:
                  type: object
                  properties:
                    bank_id: { type: integer }
                    requested: { type: integer }
                    inserted: { type: integer }
              questions: { type: array, items: { type: object } }
        400:
          description: Validation error (percentages, insufficient pool, topic not in bank)
  /tests/{test_id}/questions/ai-generate:
    post:
      tags: [AI Question Generation Flow]
      summary: Generate AI questions for review (does not import yet)
      description: |
        Stage 1 of the AI flow: generation + review storage only.
        Creates `ai_generation_requests` and `ai_generated_questions` records.
        Does NOT add questions directly to test_questions.

        Subject name and selected topic names are taken from the exam context and included in the AI prompt.
        `topic_ids` is required and every topic must belong to the exam subject.
        Provider is selected via AI_QUESTION_PROVIDER (.env):
        - `qwen` — Alibaba DashScope (DASHSCOPE_API_KEY) or HuggingFace Qwen (HF_TOKEN)
        - `gemini` — Google Gemini (GEMINI_API_KEY, default model gemini-2.5-flash)
        - `openrouter` — OpenRouter chat models (OPENROUTER_API_KEY + OPENROUTER_MODEL)
        - `huggingface` — HF router (HF_TOKEN + HF_QWEN_MODEL)
        - `placeholder` — local draft questions without external API
        - `auto` — gemini → openrouter → qwen → huggingface → placeholder (first available)
        In development, AI_FALLBACK_TO_PLACEHOLDER=true uses placeholders on quota/404 errors.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [topic_ids, count]
            properties:
              topic_ids:
                type: array
                items: { type: integer }
                minItems: 1
              count: { type: integer, example: 10 }
              type_code: { type: string, enum: [MCQ, TRUE_FALSE, MULTI_SELECT, ESSAY], default: MCQ }
              difficulty: { type: string, enum: [EASY, MEDIUM, HARD] }
              learning_objectives: { type: array, items: { type: string } }
              additional_instructions: { type: string }
      responses:
        201:
          description: AI questions generated and stored for teacher review
          schema:
            type: object
            properties:
              message: { type: string, example: "AI questions generated for review" }
              generation_request_id: { type: integer }
              generation_status: { type: string, enum: [PENDING, COMPLETED, FAILED] }
              count: { type: integer }
              subject_name: { type: string }
              ai_model: { type: string, example: Qwen/Qwen2.5-7B-Instruct }
              generated_questions: { type: array, items: { type: object } }
  /tests/{test_id}/questions/import-ai:
    post:
      tags: [AI Question Generation Flow]
      summary: Import approved AI generated questions into test
      description: |
        Stage 2 of the AI flow: import selected generated questions into the exam snapshots.
        Only questions from the specified request and in PENDING_REVIEW status are imported.
        Imported rows are marked with status IMPORTED.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [request_id, question_ids]
            properties:
              request_id: { type: integer }
              question_ids:
                type: array
                items: { type: integer }
      responses:
        201:
          description: AI questions imported into test
          schema:
            type: object
            properties:
              message: { type: string }
              count: { type: integer }
              questions: { type: array, items: { type: object } }
...
  /tests/{test_id}/questions/{test_question_id}:
    patch:
      tags: [Tests]
      summary: Update exam question snapshot
      description: |
        Partial update of a test_questions snapshot row. Allowed only while test is DRAFT
        and no student attempts exist. Does not modify the source question bank.
      security: [{ Bearer: [], WorkspaceId: [] }]
      consumes: [application/json]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - name: test_question_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          schema:
            type: object
            properties:
              type_code: { type: string, enum: [MCQ, TRUE_FALSE, MULTI_SELECT, ESSAY] }
              body: { type: string }
              image_path:
                type: string
                description: Optional image path from POST /uploads/images
              remove_image: { type: boolean, default: false }
              explanation: { type: string }
              points: { type: number }
              difficulty: { type: string, enum: [EASY, MEDIUM, HARD] }
              topic_id: { type: integer }
              choices:
                type: array
                items:
                  type: object
                  properties:
                    body: { type: string }
                    is_correct: { type: boolean }
                    order_index: { type: integer }
      responses:
        200:
          description: Question updated
        422:
          description: Test not DRAFT or validation error
    delete:
      tags: [Tests]
      summary: Remove question from exam
      description: |
        Hard-deletes the test_questions snapshot. Allowed only while test is DRAFT
        and no student attempts exist.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - name: test_question_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Question removed
        404:
          description: Question not found
        422:
          description: Test not DRAFT or attempts exist
  /tests/{test_id}/assign-students:
    post:
      tags: [Tests]
      summary: Assign students to exam whitelist
      description: |
        Creates or updates explicit allowed students for the exam.
        Each membership must be an active STUDENT enrolled in the exam subject.
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [student_membership_ids]
            properties:
              student_membership_ids:
                type: array
                items: { type: integer }
                example: [15, 22, 35, 40]
      responses:
        201:
          description: Students assigned successfully
        409:
          description: Scheduled exam time conflict for assigned students
          schema:
            $ref: "#/definitions/ScheduleConflictError"
  /tests/{test_id}/assigned-students:
    get:
      tags: [Tests]
      summary: List exam assigned students
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Assigned students list
  /tests/{test_id}/assigned-students/{membership_id}:
    delete:
      tags: [Tests]
      summary: Remove assigned student from exam
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
        - name: membership_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Student removed from assigned list
  /tests/{test_id}/publish-now:
    post:
      tags: [Tests]
      summary: Publish test immediately
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Test published
        409:
          description: Scheduled exam time conflict
          schema:
            $ref: "#/definitions/ScheduleConflictError"
...
  /tests/{test_id}/close:
    post:
      tags: [Tests]
      summary: Close test
      security: [{ Bearer: [], WorkspaceId: [] }]
      parameters:
        - name: test_id
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Test closed
...
definitions:
...
  ScheduleConflictError:
    type: object
    properties:
      error: { type: string, example: SCHEDULE_CONFLICT }
      message: { type: string }
      conflicting_test_ids:
        type: array
        items: { type: integer }
...